Added helper functions used to setup client upgrade tests
