+++ tools/test-framework/src/chain/ext/mod.rs
@@ -1,4 +1,5 @@
 pub mod bootstrap;
 pub mod fee;
 pub mod ica;
+pub mod proposal;
 pub mod transfer;
